MODINVSTOR-488: Move call number normalization tests to api package.
